get total fBEETS supply
